Fix some cleanup issues
 - php-cgi processes weren't killed when dash is the shell (eg Ubuntu 14.04)
   This is because proc_open() + proc_status() gets you the process of the
   shell instead of the executable. Same behavior with PHP5.
 - if Siege is ctrl-c'ed early, we could try to unlink an already-removed file

+++ base/Siege.php
@@ -48,7 +48,7 @@ final class Siege extends Process {
 
   public function __destruct() {
     $logfile = $this->logfile;
-    if ($logfile !== null) {
+    if ($logfile !== null && file_exists($logfile)) {
       unlink($logfile);
     }
   }
